+++ spec/presenters/activity_csv_presenter_spec.rb
@@ -1,6 +1,56 @@
 require "rails_helper"
 
 RSpec.describe ActivityCsvPresenter do
+  describe "#sector" do
+    context "when there is a non-empty sector" do
+      it "returns 'sector code: description'" do
+        activity = build(:project_activity, sector: 11110)
+        result = described_class.new(activity).sector
+        expect(result).to eq("11110: Education policy and administrative management")
+      end
+    end
+  end
+
+  describe "#aid_type" do
+    context "when there is a non-empty aid_type" do
+      it "returns 'aid type code: description'" do
+        activity = build(:project_activity, aid_type: "D02")
+        result = described_class.new(activity).aid_type
+        expect(result).to eq("D02: Other technical assistance")
+      end
+    end
+  end
+
+  describe "#flow" do
+    context "when there is a non-empty flow" do
+      it "returns 'flow code: description'" do
+        activity = build(:project_activity)
+        result = described_class.new(activity).flow
+        expect(result).to eq("10: ODA")
+      end
+    end
+  end
+
+  describe "#finance" do
+    context "when there is a non-empty finance" do
+      it "returns 'finance code: description'" do
+        activity = build(:project_activity)
+        result = described_class.new(activity).finance
+        expect(result).to eq("110: Standard grant")
+      end
+    end
+  end
+
+  describe "#tied_status" do
+    context "when there is a non-empty tied status" do
+      it "returns 'tied status code: description'" do
+        activity = build(:project_activity)
+        result = described_class.new(activity).tied_status
+        expect(result).to eq("5: Untied")
+      end
+    end
+  end
+
   describe "#benefitting_countries" do
     context "when there are benefitting countries" do
       it "returns the benefitting countries separated by semicolons" do