+++ CHANGELOG.md
@@ -1013,6 +1013,7 @@
 - Rename `activity.comments` association to better reflect its purpose of collecting all comments on an
 activity and on its child transactions (which can be actuals, refunds, and adjustments)
 - Remove Auth0 user identifier from user management
+- Add sector and sector category codes to radio button text when adding a new activity
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-102...HEAD
 [release-102]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-101...release-102